% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMSlib.R
\name{buildMSlib}
\alias{buildMSlib}
\title{Build chemical standards based MS library.}
\usage{
buildMSlib(stdInfo, stdPath, polarity)
}
\arguments{
\item{stdInfo}{A string, the path plus file name directing to a pool information file.}

\item{stdPath}{A string, the path to a directory containing the peak list files corresponding with sample pools.}

\item{polarity}{A string, the MS ionization mode.}
}
\value{
Returns the MS library object of chemical standards as specified in the input file.
}
\description{
Build chemical standards based MS library.
}
\examples{
\donttest{
testlib.neg <- buildMSlib ("tests/samplePool_neg.tsv","pool_data","negative")
testlib.pos <- buildMSlib ("tests/samplePool_pos.tsv","pool_data","positive") }
}

