% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboWeizFunctions.R
\name{runMatch}
\alias{runMatch}
\title{Match input LC-MS features to the given MS library.}
\usage{
runMatch(cameraPeaklist, rtCorrFiles, polarity, project, MSlib, massTol = 20)
}
\arguments{
\item{cameraPeaklist}{A tab delimited peak list text file, as generated by the CAMERA package.}

\item{rtCorrFiles}{The path to a directory containing the peak list files corresponding with a predefined mix of chemical standards (see below).}

\item{polarity}{The MS ionization mode (either "negative" or "positive") .}

\item{project}{Any descriptive name for the current search run.}

\item{MSlib}{MS library fileobject as generated by the \code{buildMSlibrary} function.}

\item{massTol}{The mass tolarance for mass-to-mass matches, in ppm (default is 20 ppm)}
}
\value{
NULL on a succesful completion.
}
\description{
The 'matchWeiz' multi-modular search method is applied to match the input biologically derived sample data to the 
given MS library. Retention time (RT) correction is first applied to the input peak list using a group of peak lists
originating from injections of a predefined mix of chemical standards (see \code{stdMix} data description). Several 
settings variables are used besides the input arguments, including exact mass values for possible adducts and the peak
search parameter \emph{GROUP_TH} which sets the minimum number of matches required to annotate a group of features. 
The output of the function is an R object containing all annotation data, which then needs to be converted to test 
readable format using function \code{summarizeMWresults}.
}
\examples{
\donttest{
data(sampleLib)
runMatch ("tests/Tom_all_negative_2ch_xan.tsv","tests/rt_files_neg/","negative","tom_test",MSlib.neg)
runMatch ("tests/Tom_all_negative_2ch_xan.tsv","tests/rt_files_neg/","positive","tom_test",MSlib.pos) }
}

