% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeMWresults.R
\name{summarizeMWresults}
\alias{summarizeMWresults}
\title{Summarize the \code{runMatch} annotation results}
\usage{
summarizeMWresults(project, MSlib, polarity, otherMSlib = NULL)
}
\arguments{
\item{project}{Any descriptive name for the current search run.}

\item{MSlib}{MS library object as generated by the \code{buildMSlibrary} function.}

\item{polarity}{The MS ionization mode ("negative", "positive" or "both") .}

\item{The}{MSlib MS library object corresponding with the other ionization mode.}
}
\value{
NULL on a succesful completion.
}
\description{
Summarizes the annotation data files generated by \code{runMatch}. Writes either one (single MS mode) or
two (coupled MS modes) tab delimited text files containing a summary of annotation hits sorted by annotation score.
}
\examples{
\donttest{
# A complete run using single ionization mode:
data (sampleLib)
runMatch ("tests/Tom_all_negative_2ch_xan.tsv","tests/rt_files_neg/","negative","tom_test",MSlib.neg)
summarizeMWresults ("tom_test",MSlib.neg,"negative")
# Summarizing after running \\code{runMatch} on both ionization modes:
runMatch ("tests/Tom_all_negative_2ch_xan.tsv","tests/rt_files_neg/","negative","tom_test",MSlib.neg)
runMatch ("tests/Tom_all_negative_2ch_xan.tsv","tests/rt_files_neg/","positive","tom_test",MSlib.pos)
summarizeMWresults ("tom_test",MSlib.neg,"both",MSlib.pos) }
}

