% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMSlib.R
\name{buildMSlib}
\alias{buildMSlib}
\title{Build chemical standards based MS library.}
\usage{
buildMSlib(stdInfo, stdFiles, polarity)
}
\arguments{
\item{stdInfo}{A string, the path plus file name directing to a pool information file.}

\item{stdFiles}{A string, the path to a directory containing the peak list files corresponding with sample pools.}

\item{polarity}{A string, the MS ionization mode.}
}
\value{
Returns the MS library object of chemical standards as specified in the input file.
}
\description{
Build chemical standards based MS library.
}
\examples{
\donttest{
stdInfo <- system.file("extdata/samplePool_neg.tsv",package="matchWeiz")
stdFiles <- system.file("extdata/pool_data",package="matchWeiz")
testlib.neg <- buildMSlib (stdInfo,stdFiles,"negative")
stdInfo <- system.file("extdata/samplePool_pos.tsv",package="matchWeiz")
testlib.pos <- buildMSlib (stdInfo,stdFiles,"positive") 
}
}

