% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboWeizFunctions.R
\name{runMatch}
\alias{runMatch}
\title{# Run matching and get annotations in a single ionization mode#' 
Match input LC-MS features to the given MS library.}
\usage{
runMatch(cameraPeaklist, rtCorrFiles, polarity, project, MSlib, massTol = 20)
}
\arguments{
\item{cameraPeaklist}{Path to a tab delimited peak list file, as generated by the CAMERA package.}

\item{rtCorrFiles}{Path to a directory containing peak list files corresponding with a predefined mix of chemical standards (see below).}

\item{polarity}{The MS ionization mode (either "negative" or "positive") .}

\item{project}{Any descriptive name for the current search run.}

\item{MSlib}{MS library R object as generated by the \code{buildMSlibrary} function.}

\item{massTol}{The mass tolarance for mass-to-mass matches, in ppm (default is 20 ppm)}
}
\value{
NULL on a succesful completion.
}
\description{
The 'matchWeiz' multi-modular search method is applied to match an input of biologically derived sample data to the 
given MS library. Retention time (RT) correction is first applied to the input peak list using a group of peak lists
of predefined mixes of chemical standards (see \code{stdMix} data description). Several constant variables are used 
except the input arguments, including exact mass values for possible adducts and the peak search parameter 
\emph{GROUP_TH} which sets the minimum number of matches required to annotate a group of features. These constants 
can be viewed and edited by loading the data object: \code{data("mmSettings")}. 
The output of the function is an R object containing all annotation data, which then needs to be converted to a 
readable text format using the function \code{summarizeMWresults}. Please note that two MS channels (low/high) need
to be available in the peak lists as data columns named with the suffix '01' and '02', respectively (e.g.: 'IL_140816_6701'
 and 'IL_140816_6702' would describe data of the same biological sample, but originating from channels Ms1 and Ms2).
}
\examples{
\donttest{
data(sampleLib)
tomPL_neg <- system.file("extdata/Tom_all_negative_2ch_xan.tsv",package="matchWeiz")
rtFiles_neg <- system.file("extdata/rt_files_neg",package="matchWeiz")
runMatch (tomPL_neg,rtFiles_neg,"negative","tom_test",MSlib.neg) 
}
}

