% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeMWresults.R
\name{summarizeMWresults}
\alias{summarizeMWresults}
\title{Summarizes \code{runMatch} annotation results}
\usage{
summarizeMWresults(project, MSlib, polarity, otherMSlib = NULL)
}
\arguments{
\item{project}{Any descriptive name for the current search run.}

\item{MSlib}{MS library object as generated by the \code{buildMSlibrary} function.}

\item{polarity}{The MS ionization mode ("negative", "positive" or "both") .}

\item{The}{MSlib MS library object corresponding with the other ionization mode.}
}
\value{
NULL. On succesful completion a delimited text file is written to the working directory
 containing details about the putative annotations, including: 
   compound data and scores of the individual annotation modules (columns 1-12)
   information about the corresponding library entery (columns 14-34)
   peak group data (columns 36-42) and the overall annotation score (last column).
}
\description{
Summarizes the annotation data files generated by \code{runMatch}. Writes either one (single MS mode) or
two (coupled MS modes) tab delimited text files containing a summary of annotation hits sorted by annotation score.
}
\examples{
\donttest{
data (sampleLib)
# Run matching and get annotations in a single ionization mode
tomPL_neg <- system.file("extdata/Tom_all_negative_2ch_xan.tsv",package="matchWeiz")
rtFiles_neg <- system.file("extdata/rt_files_neg",package="matchWeiz")
runMatch (tomPL_neg,rtFiles_neg,"negative","tom_test",MSlib.neg)
summarizeMWresults ("tom_test",MSlib.neg,"negative")
# Run matching and get annotations using data from both ionization mode
tomPL_pos <- system.file("extdata/Tom_all_positive_2ch_xan.tsv",package="matchWeiz")
rtFiles_pos <- system.file("extdata/rt_files_pos",package="matchWeiz")
runMatch (tomPL_pos,rtFiles_pos,"positive","tom_test",MSlib.pos)
summarizeMWresults ("tom_test",MSlib.neg,"both",MSlib.pos)
}
# Running time: the example annotation run using the demo library (total 76 enteries) 
# took 65 seconds (on an Intel i7-3820 CPU @ 3.60GHz × 8, 16Gb RAM) 
# Annotation versus the full 'WEIZMASS' library (total 3309 enteries) 
# took 341 seconds on the same hardware.
}

